refactor: migrate handler logic from handlers.py into mounts/
The 1093-line `handlers.py` and its 3222-line `test_handlers.py` have been
split into per-verb colocated modules so each mount file owns both the
handler logic and its tests.

**What moved where:**
- `handle_delete` / `make_delete_handler` → `mounts/delete.py` + `mounts/test_delete.py`
- `handle_create` / `make_create_handler` → `mounts/create.py` + `mounts/test_create.py`
- `handle_update` / `make_update_handler` → `mounts/update.py` + `mounts/test_update.py`
- `handle_get_edit_form` / `make_edit_form_handler` → `mounts/form.py` + `mounts/test_form.py`
- `handle_get_new_form` / `make_new_form_handler` → `mounts/form.py` + `mounts/test_form.py`
- `handle_detail` / `make_detail_handler` → `mounts/detail.py` + `mounts/test_detail.py`
- `handle_list` / `make_list_handler` → `mounts/list_.py` + `mounts/test_list.py`

**Shared test infrastructure** extracted to `mounts/conftest.py`
(FakeRepo, FakeAuditRepo, FixtureRow, make_user, make_audit, spec helpers).

**`handlers.py`** reduced to a re-export shim:
- Keeps _FactoryShape, _*_SHAPE constants, _param, _make_factory_handler
  (shared factory infrastructure imported lazily by each mount's make_*_handler)
- Re-exports all public handler symbols via __getattr__ so existing callers keep
  working without changes

**`_common.py`** updated: `owned_factory_makers()` now imports from individual
mount files; `assert_kind_lock()` extracted here as shared utility.

**`test_handlers.py`** reduced to 2 state-axis wrapper tests that exercise
`_wrap_state_axis_with_self_guard` from `resource_routes.py`.

**Monkeypatch target fix**: `test_clinicians.py` pagination tests patched
`handlers.DEFAULT_PAGE_SIZE`; the constant lives in `mounts/list_.py` at
runtime, so the patch target is updated to match.

No logic changes — pure reorganization. All 1997 tests pass.
